requirements.txt
Django==1.7
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.6.0
Pygments==1.6
Sphinx==1.2.1
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.18
celery==3.1.16
coverage==3.6
django-appconf==0.6
django-bootstrap3==4.7.0
django-braces==1.2.2
django-celery==3.1.16
django-countries==2.1.2
django-discover-runner==0.4
django-model-utils==1.5.0
django-rest-swagger==0.1.14
django-staticfiles==1.2.1
django-tinymce==1.5.2
djangorestframework==2.4.3
docutils==0.11
gunicorn==19.1.1
httplib2==0.9
kombu==3.0.23
logutils==0.3.3
psycopg2==2.5.3
python-dateutil==2.2
python-magic==0.4.6
pytz==2014.7
redis==2.10.1
six==1.7.2
twilio==3.6.6
wsgiref==0.1.2
